[PyTorch] Separate out pytorch_jni into pytorch_jni_jit and pytorch_jni_common.
As for internal build, we will use mobile interpreter instead of full jit, so we will need to separate the existing pytorch_jni.cpp into pytorch_jni_jit.cpp and pytorch_jni_common.cpp. pytorch_jni_common.cpp will be used both from pytorch_jni_jit.cpp(open_source) and future pytorch_jni_lite.cpp(internal).

Differential Revision: [D18387579](https://our.internmc.facebook.com/intern/diff/D18387579/)

[ghstack-poisoned]
